+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **[0.42.8] – 2026-XX-XX**  
 - Solarprognose.de entfernt, da es den Dienst leider nicht mehr gibt.  
+- api.solcast.com.au: Bei ERROR 429 zweiter Versuch nach 30 Sekunden.  
 
 **[0.42.7] – 2026-06-27**  
 - FIX: Reset-Button Wallbox-TAB.